Before Update Max Weight Count

+++ Include/Component/Animator.h
@@ -29,17 +29,17 @@ private:
 
 	AnimationClip* clip;
 
-	UINT FindScaling(float AnimationTime, const aiNodeAnim* pNodeAnim);
-	UINT FindRotation(float AnimationTime, const aiNodeAnim* pNodeAnim);
-	UINT FindPosition(float AnimationTime, const aiNodeAnim* pNodeAnim);
+	UINT FindScaling(double AnimationTime, const aiNodeAnim* pNodeAnim);
+	UINT FindRotation(double AnimationTime, const aiNodeAnim* pNodeAnim);
+	UINT FindPosition(double AnimationTime, const aiNodeAnim* pNodeAnim);
 
-	void CalcInterpolatedScaling(aiVector3D& Out, float AnimationTime, const aiNodeAnim* pNodeAnim);
-	void CalcInterpolatedRotation(aiQuaternion& Out, float AnimationTime, const aiNodeAnim* pNodeAnim);
-	void CalcInterpolatedPosition(aiVector3D& Out, float AnimationTime, const aiNodeAnim* pNodeAnim);
+	void CalcInterpolatedScaling(aiVector3D& Out, double AnimationTime, const aiNodeAnim* pNodeAnim);
+	void CalcInterpolatedRotation(aiQuaternion& Out, double AnimationTime, const aiNodeAnim* pNodeAnim);
+	void CalcInterpolatedPosition(aiVector3D& Out, double AnimationTime, const aiNodeAnim* pNodeAnim);
 
 	const aiNodeAnim* FindNodeAnim(const aiAnimation* pAnimation, const std::string NodeName);
 
-	void ReadNodeHeirarchy(float AnimationTime, const aiNode* pNode, const aiMatrix4x4& ParentTransform);
+	void ReadNodeHeirarchy(double AnimationTime, const aiNode* pNode, const aiMatrix4x4& ParentTransform);
 
 	void Play(const aiNode * _rootNode);
 };
